requirements.txt
asgiref==3.7.2
Django==5.0.2
django-ratelimit==4.1.0
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
